requirements.txt
# Landing Redesign Assistant - Dependencies
# Python 3.12+

# LLM Provider
gigachat>=0.1.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Retry logic
tenacity>=8.2.0

# Web framework (for future web interface)
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
mypy>=1.8.0
black>=24.0.0
flake8>=7.0.0

# Type stubs
types-requests>=2.31.0
types-beautifulsoup4>=4.12.0

